tests/scripts/requirements.txt
libusb1==3.1.0
pyusb==1.2.1
